+++ NEWS.md
@@ -44,6 +44,7 @@
 - Default rasterization limit in `DimPlot()` and `FeaturePlot()` changed from 50,000 to 100,000
 - `SCTransform()` now returns a formalized `Assay` subclass `SCTAssay()`
 - When using `normalization.method='SCT'` in `FindTransferAnchors()`, normalize query using reference SCT model when possible.
+- Change default Neighbor name in `FindNeighbors` to `Assay.nn`
 
 ## Removed
 - `CreateGeneActivityMatrix` replaced by `Signac::GeneActivity()`